feat: refactor ShoppingCartPage to streamline delivery method selection and enhance checkout process
- Removed DeliveryMethodSection and integrated delivery method selection directly into CheckoutSection.
- Introduced DeliverySelect component for improved delivery method options.
- Added hooks for calculating total cart price and shipping fees.
- Updated CheckoutSection to handle product purchase with selected delivery method and total price calculation.
