+++ src/pages/ShoppingCartPage/components/DeliverySelect.tsx
@@ -1,0 +1,87 @@
+import { PriceDisplay } from '@/components/PriceDisplay';
+import type { ShippingMethodType } from '@/models/grade';
+import { SHIPPING_METHOD_TYPE } from '@/models/grade';
+import { Text } from '@/ui-lib';
+import { DeliveryIcon, RocketIcon } from '@/ui-lib/components/icons';
+import { Flex, Stack } from 'styled-system/jsx';
+import { useShippingFee } from '../hooks/useShippingFee';
+
+interface DeliverySelectProps {
+  value: ShippingMethodType;
+  onChange: (value: ShippingMethodType) => void;
+}
+
+function DeliverySelect({ value, onChange }: DeliverySelectProps) {
+  const shippingFee = useShippingFee();
+
+  return (
+    <Stack gap={4}>
+      <DeliveryOption
+        title="Express"
+        description="2-3일 후 도착 예정"
+        icon={<DeliveryIcon size={28} />}
+        price={shippingFee[SHIPPING_METHOD_TYPE.EXPRESS]}
+        isSelected={value === SHIPPING_METHOD_TYPE.EXPRESS}
+        onClick={() => onChange(SHIPPING_METHOD_TYPE.EXPRESS)}
+      />
+      <DeliveryOption
+        title="Premium"
+        description="당일 배송"
+        icon={<RocketIcon size={28} />}
+        price={shippingFee[SHIPPING_METHOD_TYPE.PREMIUM]}
+        isSelected={value === SHIPPING_METHOD_TYPE.PREMIUM}
+        onClick={() => onChange(SHIPPING_METHOD_TYPE.PREMIUM)}
+      />
+    </Stack>
+  );
+}
+
+export default DeliverySelect;
+
+function DeliveryOption({
+  title,
+  description,
+  icon,
+  price,
+  isSelected,
+  onClick,
+}: {
+  title: string;
+  description: string;
+  icon: React.ReactNode;
+  price: number;
+  isSelected: boolean;
+  onClick: () => void;
+}) {
+  return (
+    <Flex
+      gap={3}
+      css={{
+        alignItems: 'center',
+        p: 5,
+        py: 4,
+        bgColor: isSelected ? 'primary.01_primary' : 'background.02_light-gray',
+        transition: 'background-color 0.3s ease',
+        rounded: '2xl',
+        color: isSelected ? 'neutral.05_white' : 'neutral.01_black',
+        cursor: 'pointer',
+      }}
+      role="button"
+      onClick={onClick}
+    >
+      {icon}
+
+      <Flex flexDir="column" gap={1} flex={1}>
+        <Text variant="B2_Regular" fontWeight={'semibold'} color={isSelected ? 'neutral.05_white' : 'neutral.01_black'}>
+          {title}
+        </Text>
+        <Text variant="C2_Medium" color={isSelected ? 'neutral.05_white' : 'neutral.02_gray'}>
+          {description}
+        </Text>
+      </Flex>
+      <Text variant="B2_Medium" fontWeight={'semibold'} color={isSelected ? 'neutral.05_white' : 'neutral.01_black'}>
+        {price ? <PriceDisplay price={price} /> : '무료'}
+      </Text>
+    </Flex>
+  );
+}
